Early functional commit
statement-scraper.py can now scrape successfully for both trades and
dividends. It needs some work to maybe enable better command-line
arguments, and potentially directory globbing for record locations, but
is coming along well.

Also, on the longer-term TODO front, I'd like to make a Portfolio class
that can be used to, say retrieve the value of a portfolio on a given
date as well as to calculate overall performance.